added ability to filter on suffix

+++ send_dir_to_s3.py
@@ -1,6 +1,6 @@
 import boto3
 import json
-import os, sys
+import os, sys, re
 import urllib, argparse
 
 class Logger:
@@ -27,7 +27,7 @@ class VerboseLogger(Logger):
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--source-directory', dest='source_directory', required=True)
-# filename suffix pattern for limiting?
+parser.add_argument('--suffix', dest='suffix', required=False, default='')
 parser.add_argument('--s3bucket', dest='s3bucket', required=True)
 parser.add_argument('--s3prefix', dest='s3prefix', required=False, default='')
 parser.add_argument('--delete-after-upload', action='store_true', default=False)
@@ -38,6 +38,7 @@ args = parser.parse_args()
 source_directory = args.source_directory
 s3bucket = args.s3bucket
 s3prefix = args.s3prefix
+suffix = args.suffix
 delete_after_upload = args.delete_after_upload
 verbose = args.verbose
 
@@ -45,8 +46,14 @@ logger = VerboseLogger() if verbose else SilentLogger()
 
 logger.log(args)
 
-# create as an array because we will be deleting as we go along
-filenames_iter = [ entry.name for entry in os.scandir(source_directory) ]
+if (suffix): # filter to only files with a matching suffix
+    pattern = f'.*{suffix}$'  # pattern for file suffix
+else:
+    pattern = '.*'            # match everything
+filter = re.compile(pattern, flags=re.IGNORECASE)
+
+# Create as an array because we will be deleting as we go along.
+filenames_iter = [ entry.name for entry in os.scandir(source_directory) if filter.match(entry.name) ]
 filenames = [*filenames_iter]
 
 if (not filenames):   # nothing to do
@@ -77,23 +84,33 @@ for filename in filenames:
 
     tagging = urllib.parse.urlencode(tags)
 
+    uploaded = False
     try:
         logger.log(f'uploading s3://{s3bucket}/{key}')
+
+        # https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html#S3.Client.put_object
+        #
+        # other params to think about:
+        #    StorageClass, ACL, encryption
+        #
         response = s3_client.put_object(Bucket=s3bucket,
                                         Key=key,
                                         Body=open(absolute_path, "rb"),
                                         Tagging=tagging)
-
         # will only get here if put_object succeeded
+        uploaded=True
         logger.log(response)
-        if (delete_after_upload):
-            os.remove(absolute_path)
-            logger.log(f'deleted {absolute_path}')
 
     except s3_client.exceptions.NoSuchBucket as e:
         print(e)
         print('no bucket named ' + event['destination'])
         # pass
+
+    if (uploaded):
+        if (delete_after_upload):
+            os.remove(absolute_path)
+            logger.log(f'deleted {absolute_path}')
+
